Structuring streaming-middleware to its own folder
